Move test to the razorpageswebsite to avoid conflicts
